# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentVerse API",
    "version": "0.1.0-alpha"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/integrations/catalog": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List Catalog Route",
        "description": "The marketplace.\n\nStill workspace-authenticated even though the catalog is\nplatform-wide: browsing it reveals what the platform integrates with,\nwhich is not information for anonymous callers.",
        "operationId": "list_catalog_route_api_v1_workspaces__workspace_id__integrations_catalog_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/McpServerResponse"
                  },
                  "title": "Response List Catalog Route Api V1 Workspaces  Workspace Id  Integrations Catalog Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/integrations": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Install From Catalog Route",
        "description": "Installs a vetted catalog entry.\n\nAdmin-gated: an install decides which third party the workspace's\nagents may reach, and its command (for stdio) runs on the worker\nfleet.",
        "operationId": "install_from_catalog_route_api_v1_workspaces__workspace_id__integrations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallFromCatalogRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstalledServerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List Installed Route",
        "operationId": "list_installed_route_api_v1_workspaces__workspace_id__integrations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstalledServerResponse"
                  },
                  "title": "Response List Installed Route Api V1 Workspaces  Workspace Id  Integrations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/integrations/custom": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Register Custom Server Route",
        "description": "Registers a user's own MCP endpoint.\n\n`mcp_server_id` is null, which is what marks this as custom \u2014 and\nwhat makes `factory._build_stdio` refuse it, since stdio is available\nonly to vetted catalog entries.",
        "operationId": "register_custom_server_route_api_v1_workspaces__workspace_id__integrations_custom_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCustomServerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstalledServerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/integrations/{installed_server_id}/credentials": {
      "put": {
        "tags": [
          "integrations"
        ],
        "summary": "Put Credential Route",
        "description": "Writes or rotates a credential.\n\nThe value is encrypted here and never stored, logged, or returned in\nplaintext. The AAD binds the ciphertext to this\n`(workspace, server, key)` \u2014 a row copied into another workspace\nfails to decrypt rather than handing over a working credential.",
        "operationId": "put_credential_route_api_v1_workspaces__workspace_id__integrations__installed_server_id__credentials_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "installed_server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Installed Server Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutCredentialRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List Credentials Route",
        "description": "Which credentials exist, and when they were last rotated.\n\nNever their values \u2014 `CredentialResponse` has no field for one. There\nis deliberately no `GET /{key}/value` route: building the read path\nat all is what later gets loosened.",
        "operationId": "list_credentials_route_api_v1_workspaces__workspace_id__integrations__installed_server_id__credentials_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "installed_server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Installed Server Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CredentialResponse"
                  },
                  "title": "Response List Credentials Route Api V1 Workspaces  Workspace Id  Integrations  Installed Server Id  Credentials Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "McpServerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "transport": {
            "type": "string",
            "enum": [
              "stdio",
              "sse",
              "streamable_http"
            ],
            "title": "Transport"
          },
          "availability": {
            "type": "string",
            "enum": [
              "official",
              "community",
              "custom_required"
            ],
            "title": "Availability"
          },
          "auth_scheme": {
            "type": "string",
            "enum": [
              "none",
              "api_key",
              "bearer_token",
              "basic",
              "oauth2",
              "jwt",
              "custom_header"
            ],
            "title": "Auth Scheme"
          },
          "required_credentials": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Credentials"
          },
          "oauth_scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Oauth Scopes"
          },
          "documentation_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentation Url"
          },
          "icon_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Slug"
          },
          "is_installable": {
            "type": "boolean",
            "title": "Is Installable"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "category",
          "transport",
          "availability",
          "auth_scheme",
          "required_credentials",
          "oauth_scopes",
          "documentation_url",
          "icon_slug",
          "is_installable"
        ],
        "title": "McpServerResponse",
        "description": "A marketplace catalog entry."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}